[COZY-139] refactor: 클래스명 변경 및 패키지 위치 이동
