+++ src/main/java/com/cozymate/cozymate_server/global/fcm/NotificationSender.java
@@ -1,13 +1,13 @@
-package com.cozymate.cozymate_server.domain.notificationlog.firebase;
+package com.cozymate.cozymate_server.global.fcm;
 
 import com.cozymate.cozymate_server.domain.member.Member;
 import com.cozymate.cozymate_server.domain.member.MemberRepository;
 import com.cozymate.cozymate_server.domain.notificationlog.NotificationLog;
 import com.cozymate.cozymate_server.domain.notificationlog.enums.NotificationType;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.GroupTargetVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.OneTargetReverseVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.OneTargetVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.TwoTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.GroupTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.OneTargetReverseVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.OneTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.TwoTargetVO;
 import com.cozymate.cozymate_server.domain.notificationlog.service.NotificationLogCommandService;
 import com.cozymate.cozymate_server.global.response.code.status.ErrorStatus;
 import com.cozymate.cozymate_server.global.response.exception.GeneralException;
