+++ src/test/java/com/cozymate/cozymate_server/domain/notificationlog/firebase/NotificationSenderTest.java
@@ -3,10 +3,12 @@ package com.cozymate.cozymate_server.domain.notificationlog.firebase;
 import com.cozymate.cozymate_server.domain.member.Member;
 import com.cozymate.cozymate_server.domain.member.MemberRepository;
 import com.cozymate.cozymate_server.domain.notificationlog.NotificationLog;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.GroupTargetVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.OneTargetReverseVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.OneTargetVO;
-import com.cozymate.cozymate_server.domain.notificationlog.firebase.NotificationTargetVO.TwoTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationContentVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationSender;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.GroupTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.OneTargetReverseVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.OneTargetVO;
+import com.cozymate.cozymate_server.global.fcm.NotificationTargetVO.TwoTargetVO;
 import com.cozymate.cozymate_server.domain.notificationlog.service.NotificationLogCommandService;
 import com.cozymate.cozymate_server.global.response.exception.GeneralException;
 import com.google.firebase.messaging.FirebaseMessaging;
